Encapsulated service documant parsing
