+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/dccticketing/core/server/DccTicketingServer.kt
@@ -1,13 +1,9 @@
 package de.rki.coronawarnapp.dccticketing.core.server
 
-import com.google.gson.Gson
 import dagger.Lazy
 import dagger.Reusable
-import de.rki.coronawarnapp.dccticketing.core.allowlist.data.DccTicketingValidationServiceAllowListEntry
-import de.rki.coronawarnapp.dccticketing.core.check.DccTicketingServerCertificateCheckException
 import de.rki.coronawarnapp.dccticketing.core.check.DccTicketingServerCertificateChecker
 import de.rki.coronawarnapp.dccticketing.core.transaction.DccJWK
-import de.rki.coronawarnapp.dccticketing.core.transaction.DccTicketingServiceIdentityDocument
 import de.rki.coronawarnapp.dccticketing.core.server.DccTicketingServerException.ErrorCode
 import de.rki.coronawarnapp.exception.http.CwaClientError
 import de.rki.coronawarnapp.exception.http.CwaUnknownHostException
@@ -15,8 +11,6 @@ import de.rki.coronawarnapp.exception.http.NetworkConnectTimeoutException
 import de.rki.coronawarnapp.exception.http.NetworkReadTimeoutException
 import de.rki.coronawarnapp.tag
 import de.rki.coronawarnapp.util.coroutine.DispatcherProvider
-import de.rki.coronawarnapp.util.serialization.BaseGson
-import de.rki.coronawarnapp.util.serialization.fromJson
 import kotlinx.coroutines.withContext
 import okhttp3.ResponseBody
 import retrofit2.Response
@@ -27,7 +21,6 @@ import javax.inject.Inject
 class DccTicketingServer @Inject constructor(
     private val dccTicketingApiV1Lazy: Lazy<DccTicketingApiV1>,
     private val dispatcherProvider: DispatcherProvider,
-    @BaseGson private val gson: Gson,
     private val serverCertificateChecker: DccTicketingServerCertificateChecker
 ) {
 
@@ -37,21 +30,9 @@ class DccTicketingServer @Inject constructor(
     @Throws(DccTicketingServerException::class)
     suspend fun getServiceIdentityDocument(
         url: String
-    ): DccTicketingServiceIdentityDocument = withContext(dispatcherProvider.IO) {
+    ): Response<ResponseBody> = withContext(dispatcherProvider.IO) {
         Timber.tag(TAG).d("getServiceIdentityDocument(url=%s)", url)
-        get(url).parse()
-    }
-
-    @Throws(DccTicketingServerException::class, DccTicketingServerCertificateCheckException::class)
-    suspend fun getServiceIdentityDocumentAndValidateServerCert(
-        url: String,
-        allowList: Set<DccTicketingValidationServiceAllowListEntry>
-    ): DccTicketingServiceIdentityDocument = withContext(dispatcherProvider.IO) {
-        Timber.tag(TAG).d("getServiceIdentityDocument(url=%s, allowlist=%s)", url, allowList)
-        get(url).run {
-            validateAgainstAllowlist(allowList = allowList)
-            parse()
-        }
+        get(url)
     }
 
     private suspend fun get(url: String): Response<ResponseBody> = try {
@@ -67,27 +48,9 @@ class DccTicketingServer @Inject constructor(
         }.let { DccTicketingServerException(errorCode = it, cause = e) }
     }
 
-    private inline fun <reified T> Response<ResponseBody>.parse(): T = try {
-        Timber.tag(TAG).d("Parsing response=%s", this)
-        body()!!.charStream().use { gson.fromJson(it) }
-    } catch (e: Exception) {
-        Timber.e(e, "Parsing failed")
-        throw DccTicketingServerException(errorCode = ErrorCode.PARSE_ERR, cause = e)
-    }
-
     private fun Response<ResponseBody>.validateAgainstJwkSet(jwkSet: Set<DccJWK>) {
         Timber.tag(TAG).d("Validating response with jwk set=%s", jwkSet)
         serverCertificateChecker.checkCertificate(raw(), jwkSet)
-    }
-
-    private fun Response<ResponseBody>.validateAgainstAllowlist(
-        allowList: Set<DccTicketingValidationServiceAllowListEntry>
-    ) {
-        Timber.tag(TAG).d("Validating response with against allow list=%s", allowList)
-        serverCertificateChecker.checkCertificateAgainstAllowlist(
-            response = raw(),
-            allowlist = allowList
-        )
     }
 
     @Suppress("BlockingMethodInNonBlockingContext")
